requirements.txt
argparse
progressbar2